The Resume Builder App is a frontend web application designed to help users create professional resumes easily and efficiently. The app is built using React and Redux, two popular JavaScript libraries, and allows users to select from a variety of resume templates to customize their own resume.
The project contains four main pages:-
=>Home page:
 In this ,on the top, there is a navigation bar  with logo and showing three links at the  right to navigate -
->Resume Template- it will take the user to the home page
->My Resume- here the user can see their latest created resume
->About Us

And below of the navbar, it displays four predefined templates for the resumes,
When the users hover over any of the templates they will see the ‘use Template’ button and on clicking on it, it will navigate them to the details filling page.

=>Details Filling page:
On this page, there is SideNavigationBar in the left section and in the right section, by default the personal info section will be rendered.
But the user can navigate to other pages either using the sidebar or clicking the next button.
Like if the user clicks work experience tab, it would change the filling layout from personal Info to work experience, same way education and key skills details filling layout will render on the details filling page.

=>My Resume Page:
When the user is done with filling the details and clicks on the preview button ,
It will direct the user to this page, where the user created resume will be shown.
And at the top of the page, there are two buttons named as ‘Go-Back’, which will direct the user to the Key Skills filling layout of the details filling page and the other is ‘Save Resume’ , when the user clicks  it , it will save the created resume in the user’s computer memory and after the download is successful , the modal will pop-up showing the success message.

⇒ About Us Page: this page will provide general information about the application.

The Resume Builder App offers the following features:
User-friendly interface: The app has an intuitive and easy-to-use interface, making it accessible to users of all skill levels.
Multiple templates: Users can choose from a variety of templates to create a personalized and professional-looking resume.
Customization options: The app allows users to customize their resumes by adding or removing sections, changing fonts and colors, and more.
Easy editing: The app allows users to edit their resumes easily and make changes as needed.
Download and print options: Once a user has completed their resume, they can download it as a PDF or print it directly from the app.
To use the Resume Builder App, simply follow these steps:
Go to the app's website and select a template that you like.
Add your personal information, work experience, education and keyskills to the template.
Preview your resume to ensure that it looks professional and well-designed.
Download your resume as a PDF or print it directly from the app.

The Resume Builder App was built using the following technologies:
React: A JavaScript library for building user interfaces.
Redux: A predictable state container for JavaScript apps.
HTML/CSS: The standard markup language and styling language for creating web pages.